requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
colorama==0.4.4
Django==4.0
django-qrcode==0.3
et-xmlfile==1.1.0
imageio==2.13.5
MyQR==2.3.1
numpy==1.22.0
openpyxl==3.0.9
pandas==1.3.5
Pillow==9.0.0
python-dateutil==2.8.2
pytz==2021.3
qrcode==7.3.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
django-import-export
